AJS-61 moved tests back to the "tests" folder
