--- tests/browser.chrome.conf.js
@@ -1,19 +1,0 @@
-var sharedConfig = require('./karma.conf.js');
-
-module.exports = function(config) {
-
-  sharedConfig(config);
-
-  config.plugins.push('karma-chrome-launcher');
-
-  // you can define custom flags
-  config.customLaunchers = {
-    ChromeUM: {
-      base: 'Chrome',
-      flags: ['--use-fake-ui-for-media-stream']
-    }
-  };
-
-  // Run each browser manually for automated test
-  config.browsers = ['ChromeUM'];
-};
