+++ tests/unit/MediaStream.event.spec.js
@@ -26,19 +26,10 @@ describe('MediaStream | EventHandler', function() {
 	/* WebRTC Object should be initialized in Safari/IE Plugin */
 	before(function (done) {
 		this.timeout(testItemTimeout);
-
-		if (window.webrtcDetectedBrowser !== 'IE' && window.webrtcDetectedBrowser !== 'Safari') {
-			AdapterJS.onwebrtcreadyDone = true;
-		}
-
-		if (!AdapterJS.onwebrtcreadyDone) {
-			AdapterJS.onwebrtcready = function () {
-				done();
-			};
-
-		} else {
-			done();
-		}
+    
+    AdapterJS.webRTCReady(function() {
+    	done();
+    });
 	});
 
 	/* Get User Media */
@@ -48,13 +39,10 @@ describe('MediaStream | EventHandler', function() {
 		window.navigator.getUserMedia({
 			audio: true,
 			video: true
-
 		}, function (data) {
-
 			stream = data;
-			track = data.polygetAudioTracks()[0];
+			track = data.getAudioTracks()[0];
 			done();
-
 		}, function (error) {
 			throw error;
 		});
@@ -64,50 +52,50 @@ describe('MediaStream | EventHandler', function() {
 		this.timeout(testItemTimeout);
 
 		stream.onremovetrack = function () {
+			//Note(J-O): doens't pass on either Chrome or the plugin
 		  done();
 		};
 
-		stream.polyremoveTrack(track);
+		stream.removeTrack(track);
 	});
 
 	it('MediaStream.onaddtrack :: emit', function (done) {
 		this.timeout(testItemTimeout);
 
 		stream.onaddtrack = function () {
+			//Note(J-O): doens't pass on either Chrome or the plugin
 		  done();
 		};
 
-		stream.polyaddTrack(track);
+		stream.addTrack(track);
 	});
 
-	it('MediaStream.onended :: emit < When > MediaStream.polystop()', function (done) {
+	it('MediaStream.onended :: emit < When > MediaStream.stop()', function (done) {
 		this.timeout(testItemTimeout);
 
 		stream.onended = function () {
 		  done();
 		};
 
-		stream.polystop();
+		stream.stop();
 	});
 
-	it('MediaStream.onended :: emit < When > for MediaStreamTrack in MediaStream >> MediaStreamTrack.polystop()', function (done) {
+	it('MediaStream.onended :: emit < When > for MediaStreamTrack in MediaStream >> MediaStreamTrack.stop()', function (done) {
 		this.timeout(testItemTimeout);
 
 		stream.onended = function () {
 		  done();
 		};
 
-		var i, j;
+		var audioTracks = stream.getAudioTracks();
+		var videoTracks = stream.getVideoTracks();
 
-		var audioTracks = stream.polygetAudioTracks();
-		var videoTracks = stream.polygetVideoTracks();
-
-		for (i = 0; i < audioTracks.length; i += 1) {
-			audioTracks[i].polystop();
+		for (var i = 0; i < audioTracks.length; i += 1) {
+			audioTracks[i].stop();
 		}
 
-		for (j = 0; j < videoTracks.length; j += 1) {
-			videoTracks[j].polystop();
+		for (var j = 0; j < videoTracks.length; j += 1) {
+			videoTracks[j].stop();
 		}
 	});
 });
