+++ tests/unit/MediaStreamError.spec.js
@@ -23,18 +23,9 @@ describe('MediaStreamError', function() {
   before(function (done) {
     this.timeout(testItemTimeout);
 
-    if (window.webrtcDetectedBrowser !== 'IE' && window.webrtcDetectedBrowser !== 'Safari') {
-      AdapterJS.onwebrtcreadyDone = true;
-    }
-
-    if (!AdapterJS.onwebrtcreadyDone) {
-      AdapterJS.onwebrtcready = function () {
-        done();
-      };
-
-    } else {
-      done();
-    }
+	AdapterJS.webRTCReady(function() {
+		done();
+	});
   });
 
 	it('MediaStreamError.name === "NotSupportedError" < When > MediaStreamConstraints === {}', function (done) {
