+++ tests/unit/MediaStreamTrack.event.spec.js
@@ -26,18 +26,9 @@ describe('MediaStreamTrack | EventHandler', function() {
 	before(function (done) {
 		this.timeout(testItemTimeout);
 
-		if (window.webrtcDetectedBrowser !== 'IE' && window.webrtcDetectedBrowser !== 'Safari') {
-			AdapterJS.onwebrtcreadyDone = true;
-		}
-
-		if (!AdapterJS.onwebrtcreadyDone) {
-			AdapterJS.onwebrtcready = function () {
-				done();
-			};
-
-		} else {
+		AdapterJS.webRTCReady(function() {
 			done();
-		}
+		});
 	});
 
 	/* Get User Media */
@@ -51,7 +42,7 @@ describe('MediaStreamTrack | EventHandler', function() {
 
 		}, function (data) {
 			stream = data;
-			track = data.polygetAudioTracks()[0];
+			track = data.getAudioTracks()[0];
 
 			done();
 
@@ -68,7 +59,7 @@ describe('MediaStreamTrack | EventHandler', function() {
 
 	it.skip('MediaStreamTrack.onoverconstrained :: emit', function () {});
 
-	it('MediaStreamTrack.onended :: emit < When > MediaStreamTrack.polystop()', function (done) {
+	it('MediaStreamTrack.onended :: emit < When > MediaStreamTrack.stop()', function (done) {
 		this.timeout(testItemTimeout);
 
 		var hasTriggered = false;
@@ -77,10 +68,10 @@ describe('MediaStreamTrack | EventHandler', function() {
 		  done();
 		};
 
-		track.polystop();
+		track.stop();
 	});
 
-	it('MediaStreamTrack.onended :: emit < When > MediaStream.polystop()', function (done) {
+	it('MediaStreamTrack.onended :: emit < When > MediaStream.stop()', function (done) {
 		this.timeout(testItemTimeout);
 
 		var i, j;
@@ -88,8 +79,8 @@ describe('MediaStreamTrack | EventHandler', function() {
 		var audioEndedTriggered = 0;
 		var videoEndedTriggered = 0;
 
-		var audioTracks = stream.polygetAudioTracks();
-		var videoTracks = stream.polygetVideoTracks();
+		var audioTracks = stream.getAudioTracks();
+		var videoTracks = stream.getVideoTracks();
 
 		var audioTriggeredFn = function () {
 			audioEndedTriggered += 1;
@@ -107,7 +98,7 @@ describe('MediaStreamTrack | EventHandler', function() {
 			videoTracks[j].onended = videoTriggeredFn;
 		}
 
-		stream.polystop();
+		stream.stop();
 
 		setTimeout(function () {
 			expect(audioEndedTriggered).to.equal(audioTracks.length);
