+++ tests/unit/RTCPeerConnection.offerconstraints.spec.js
@@ -12,7 +12,7 @@ var testTimeout = 35000;
 var gUMTimeout = 25000;
 
 // Test item timeout
-var testItemTimeout = 2000;
+var testItemTimeout = 5000;
 
 
 describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
@@ -28,18 +28,18 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
   before(function (done) {
     this.timeout(testItemTimeout);
 
-    if (window.webrtcDetectedBrowser !== 'IE' && window.webrtcDetectedBrowser !== 'Safari') {
-      AdapterJS.onwebrtcreadyDone = true;
-    }
-
-    if (!AdapterJS.onwebrtcreadyDone) {
-      AdapterJS.onwebrtcready = function () {
+    AdapterJS.webRTCReady(function() {
+      // get a reference stream for all tests
+      window.navigator.getUserMedia({
+        audio: true,
+        video: true
+      }, function (data) {
+        stream = data;
         done();
-      };
-
-    } else {
-      done();
-    }
+      }, function (error) {
+        throw error;
+      });
+    });
   });
 
   /* Create peer objects */
@@ -52,18 +52,8 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
       iceServers: []
     });
 
-    window.navigator.getUserMedia({
-      audio: true,
-      video: true
-
-    }, function (data) {
-      stream = data;
-      peer1.addStream(stream);
-      done();
-
-    }, function (error) {
-      throw error;
-    });
+    peer1.addStream(stream);
+    done();
   });
 
 
@@ -83,7 +73,7 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
       peer2.onaddstream = function (event) {
@@ -91,9 +81,11 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
         expect(remoteStream.getAudioTracks()).to.have.length(0);
         expect(remoteStream.getVideoTracks()).to.have.length(1);
+
+        done();
       };
 
-      connect(peer1, peer2, null, constraints);
+      connect(peer1, peer2, constraints);
     });
 
   })({ offerToReceiveVideo: true, offerToReceiveAudio: false });
@@ -101,17 +93,19 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
       peer2.onaddstream = function (event) {
         var remoteStream = event.stream || event;
 
-        expect(remoteStream.getAudioTracks()).to.have.length(0);
+        expect(remoteStream.getAudioTracks()).to.have.length(1);
         expect(remoteStream.getVideoTracks()).to.have.length(1);
+
+        done();
       };
 
-      connect(peer1, peer2, null, constraints);
+      connect(peer1, peer2, constraints);
     });
 
   })({ offerToReceiveVideo: true });
@@ -119,7 +113,7 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
       peer2.onaddstream = function (event) {
@@ -127,9 +121,11 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
         expect(remoteStream.getAudioTracks()).to.have.length(1);
         expect(remoteStream.getVideoTracks()).to.have.length(0);
+
+        done();
       };
 
-      connect(peer1, peer2, null, constraints);
+      connect(peer1, peer2, constraints);
     });
 
   })({ offerToReceiveVideo: false, offerToReceiveAudio: true });
@@ -137,7 +133,7 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
       peer2.onaddstream = function (event) {
@@ -145,9 +141,11 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
         expect(remoteStream.getAudioTracks()).to.have.length(1);
         expect(remoteStream.getVideoTracks()).to.have.length(0);
+
+        done();
       };
 
-      connect(peer1, peer2, null, constraints);
+      connect(peer1, peer2, constraints);
     });
 
   })({ offerToReceiveAudio: true });
@@ -155,7 +153,7 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
       peer2.onaddstream = function (event) {
@@ -163,9 +161,11 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
         expect(remoteStream.getAudioTracks()).to.have.length(1);
         expect(remoteStream.getVideoTracks()).to.have.length(1);
+
+        done();
       };
 
-      connect(peer1, peer2, null, constraints);
+      connect(peer1, peer2, constraints);
     });
 
   })({ offerToReceiveVideo: true, offerToReceiveAudio: true });
@@ -173,7 +173,7 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
       peer2.onaddstream = function (event) {
@@ -181,9 +181,11 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
         expect(remoteStream.getAudioTracks()).to.have.length(0);
         expect(remoteStream.getVideoTracks()).to.have.length(0);
+
+        done();
       };
 
-      connect(peer1, peer2, null, constraints);
+      connect(peer1, peer2, constraints);
     });
 
   })({ offerToReceiveVideo: false, offerToReceiveAudio: false });
@@ -191,12 +193,19 @@ describe('RTCPeerConnection.createOffer | RTCOfferOptions', function() {
 
   (function (constraints) {
 
-    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function () {
+    it('RTCPeerConnection.createOffer(successCb, failureCb, ' + printJSON(constraints) + ')', function (done) {
       this.timeout(testItemTimeout);
 
-      expect(function () {
-        connect(peer1, peer2, null, constraints);
-      }).to.throw(Error);
+      peer2.onaddstream = function (event) {
+        var remoteStream = event.stream || event;
+
+        expect(remoteStream.getAudioTracks()).to.have.length(1);
+        expect(remoteStream.getVideoTracks()).to.have.length(1);
+
+        done();
+      };
+
+      connect(peer1, peer2, constraints);
     });
 
   })({});
