+++ tests/unit/RTCPeerConnection.param.spec.js
@@ -22,18 +22,9 @@ describe('RTCPeerConnection | Parameters', function() {
   before(function (done) {
     this.timeout(testItemTimeout);
 
-    if (window.webrtcDetectedBrowser !== 'IE' && window.webrtcDetectedBrowser !== 'Safari') {
-      AdapterJS.onwebrtcreadyDone = true;
-    }
-
-    if (!AdapterJS.onwebrtcreadyDone) {
-      AdapterJS.onwebrtcready = function () {
-        done();
-      };
-
-    } else {
+    AdapterJS.webRTCReady(function() {
       done();
-    }
+    });
   });
 
   it('new RTCPeerConnection() -> Error', function () {
