--- tmp/test-results.testxml
@@ -1,2 +1,0 @@
-<?xml version="1.0"?>
-<Tests/>
