--- tmp/unit/RTCPeerConnection.prop.spec.js
@@ -1,661 +1,0 @@
-//mocha.bail();
-//mocha.run();
-
-var expect = chai.expect;
-var assert = chai.assert;
-var should = chai.should;
-
-// Test timeouts
-var testTimeout = 35000;
-
-// Get User Media timeout
-var gUMTimeout = 25000;
-
-// Test item timeout
-var testItemTimeout = 5000;
-
-var err = new ReferenceError('This is a bad function.');
-
-describe('RTCPeerConnection | Properties', function() {
-  this.timeout(testTimeout);
-
-  var peer1 = null;
-  var peer2 = null;
-  var stream = null;
-
-  var offer = null;
-  var answer = null;
-
-  var candidates1 = [];
-  var candidates2 = [];
-
-
-  /* WebRTC Object should be initialized in Safari/IE Plugin */
-  before(function (done) {
-    this.timeout(testItemTimeout);
-
-    var getMedia = function () {
-      window.navigator.getUserMedia({
-        audio: true,
-        video: true
-
-      }, function (data) {
-        stream = data;
-        done();
-
-      }, function (error) {
-        throw error;
-      });
-    };
-
-    if (window.webrtcDetectedBrowser !== 'IE' && window.webrtcDetectedBrowser !== 'Safari') {
-      AdapterJS.onwebrtcreadyDone = true;
-    }
-
-    if (!AdapterJS.onwebrtcreadyDone) {
-      AdapterJS.onwebrtcready = getMedia;
-
-    } else {
-      getMedia();
-    }
-  });
-
-    /* Get User Media */
-  beforeEach(function (done) {
-    this.timeout(testItemTimeout);
-
-    peer1 = new RTCPeerConnection({ iceServers: [] });
-    peer2 = new RTCPeerConnection({ iceServers: [] });
-
-    done();
-  });
-
-  it('RTCPeerConnection.localDescription :: object', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(peer1).to.have.property('localDescription');
-    expect(peer2).to.have.property('localDescription');
-    expect(peer1.localDescription).to.equal(null); // fails on Chrome
-    expect(peer2.localDescription).to.equal(null); // fails on Chrome
-
-    done();
-  });
-
-  it('RTCPeerConnection.remoteDescription :: object', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(peer1).to.have.property('remoteDescription');
-    expect(peer2).to.have.property('remoteDescription');
-    expect(peer1.remoteDescription).to.equal(null); // fails on Chrome
-    expect(peer2.remoteDescription).to.equal(null); // fails on Chrome
-
-    done();
-  });
-
-  it('RTCPeerConnection.signalingState :: string', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.typeOf(peer1.signalingState, 'string');
-    assert.typeOf(peer2.signalingState, 'string');
-    expect(peer1.signalingState).to.equal('stable');
-    expect(peer2.signalingState).to.equal('stable');
-
-    done();
-  });
-
-  it('RTCPeerConnection.iceConnectionState :: string', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.typeOf(peer1.iceConnectionState, 'string');
-    assert.typeOf(peer2.iceConnectionState, 'string');
-    expect(peer1.iceConnectionState).to.equal('new');
-    expect(peer2.iceConnectionState).to.equal('new');
-
-    done();
-  });
-
-  it('RTCPeerConnection.iceGatheringState :: string', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.typeOf(peer1.iceGatheringState, 'string');
-    assert.typeOf(peer2.iceGatheringState, 'string');
-    expect(peer1.iceGatheringState).to.equal('new');
-    expect(peer2.iceGatheringState).to.equal('new');
-
-    done();
-  });
-
-  it('RTCPeerConnection.canTrickleIceCandidates :: boolean', function (done) {
-    // Note(J-O) fails on Chrome
-
-    this.timeout(testItemTimeout);
-
-    assert.typeOf(peer1.canTrickleIceCandidates, 'boolean');
-    assert.typeOf(peer2.canTrickleIceCandidates, 'boolean');
-    expect(peer1.canTrickleIceCandidates).to.equal(true);
-    expect(peer2.canTrickleIceCandidates).to.equal(true);
-
-    done();
-  });
-
-  it('RTCPeerConnection.getLocalStreams :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.getLocalStreams, 'function');
-    assert.equal(typeof peer2.getLocalStreams, 'function');
-
-    var result1 = peer1.getLocalStreams();
-    var result2 = peer2.getLocalStreams();
-    assert.instanceOf(result1, Array);
-    assert.instanceOf(result2, Array);
-
-    done();
-  });
-
-  it('RTCPeerConnection.getRemoteStreams :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.getRemoteStreams, 'function');
-    assert.equal(typeof peer2.getRemoteStreams, 'function');
-
-    var result1 = peer1.getRemoteStreams();
-    var result2 = peer2.getRemoteStreams();
-    assert.instanceOf(result1, Array);
-    assert.instanceOf(result2, Array);
-
-    done();
-  });
-
-  it('RTCPeerConnection.addStream :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.addStream, 'function');
-    assert.equal(typeof peer2.addStream, 'function');
-
-    peer1.addStream(stream);
-    peer2.addStream(stream);
-    expect(peer1.getLocalStreams()).to.have.length(1);
-    expect(peer2.getLocalStreams()).to.have.length(1);
-
-    done();
-  });
-
-  it('RTCPeerConnection.addStream :: method < For > Multi-stream Feature', function (done) {
-    this.timeout(testItemTimeout);
-
-    peer1.addStream(stream);
-    peer2.addStream(stream);
-
-    expect(peer1.getLocalStreams()).to.have.length(1);
-    expect(peer2.getLocalStreams()).to.have.length(1);
-
-    getUserMedia({
-        audio: true,
-        video: true
-      }, function (s) {
-        peer1.addStream(s);
-        peer2.addStream(s);
-
-        expect(peer1.getLocalStreams()).to.have.length(2);
-        expect(peer2.getLocalStreams()).to.have.length(2);
-
-        done();
-      }, function (error) {
-        throw error;
-      }
-    );
-
-  });
-
-  it('RTCPeerConnection.removeStream :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.removeStream, 'function');
-    assert.equal(typeof peer2.removeStream, 'function');
-
-    peer1.removeStream(stream);
-    peer2.removeStream(stream);
-    expect(peer1.getLocalStreams()).to.have.length(0);
-    expect(peer2.getLocalStreams()).to.have.length(0);
-
-    done();
-  });
-
-  it('RTCPeerConnection.getConfiguration :: method', function (done) {
-    // Note(J-O) fails on Chrome
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.getConfiguration, 'function');
-    assert.equal(typeof peer2.getConfiguration, 'function');
-
-    var result1 = peer1.getConfiguration();
-    var result2 = peer1.getConfiguration();
-    assert.typeOf(result1, 'object');
-    assert.typeOf(result2, 'object');
-
-    done();
-  });
-
-  it.skip('RTCPeerConnection.updateIce :: method', function (done) {});
-
-  it('RTCPeerConnection.getStreamById :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.getStreamById, 'function');
-    assert.equal(typeof peer2.getStreamById, 'function');
-
-    var result1 = peer1.getStreamById(stream.id);
-    var result2 = peer2.getStreamById(stream.id);
-
-    expect(result1).to.be.null;
-    expect(result2).to.be.null;
-
-    peer1.addStream(stream);
-    peer2.addStream(stream);
-
-    var result1 = peer1.getStreamById(stream.id);
-    var result2 = peer2.getStreamById(stream.id);
-
-    expect(result1).to.equal(stream);
-    expect(result2).to.equal(stream);
-
-    done();
-  });
-
-  it('RTCPeerConnection.createDataChannel :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.createDataChannel, 'function');
-    assert.equal(typeof peer2.createDataChannel, 'function');
-
-    var result1 = peer1.createDataChannel('Label');
-    var result2 = peer2.createDataChannel('Label2');
-
-    assert.typeOf(result1, 'object');
-    assert.typeOf(result2, 'object');
-
-    var result3 = peer1.createDataChannel('Label');
-    var result4 = peer2.createDataChannel('Label2');
-
-    assert.typeOf(result3, 'object');
-    assert.typeOf(result4, 'object');
-
-    done();
-  });
-
-  it('RTCPeerConnection.createDTMFSender :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.createDTMFSender, 'function');
-    assert.equal(typeof peer2.createDTMFSender, 'function');
-
-    var track = stream.getAudioTracks()[0];
-
-    var result1 = peer1.createDTMFSender(track);
-    var result2 = peer2.createDTMFSender(track);
-
-    assert.typeOf(result1, 'object');
-    assert.typeOf(result2, 'object');
-
-    expect(result1.track).to.equal(track);
-    expect(result2.track).to.equal(track);
-
-    done();
-  });
-
-  it('RTCPeerConnection.createOffer :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.createOffer, 'function');
-    assert.equal(typeof peer2.createOffer, 'function');
-
-    peer1.addStream(stream);
-    peer2.addStream(stream);
-
-    peer1.createOffer(function (sdp) {
-
-      offer = sdp;
-
-      expect(sdp.type).to.equal('offer');
-      assert.typeOf(sdp.sdp, 'string');
-
-      expect(sdp.sdp).to.contain('=audio');
-      expect(sdp.sdp).to.contain('=video');
-      // expect(sdp.sdp).to.contain('=application'); // Note(J-O): I don't know what this is, commented
-
-      done();
-
-    }, function (error) {
-      throw error;
-
-    }, {
-      offerToReceiveAudio: true,
-      offerToReceiveVideo: true
-    });
-  });
-
-  it('RTCPeerConnection.createOffer :: method < When > RTCPeerConnection.createOffer(invalid parameters)', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(function () {
-      peer1.createOffer(function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.createOffer(function () {});
-    }).to.throw(Error);
-
-    done();
-  });
-
-  it('RTCPeerConnection.setRemoteDescription(offer) :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer2.setRemoteDescription, 'function');
-
-    if (offer === null) {
-      throw new Error('Offer Session Description is not yet created');
-    }
-
-    peer2.setRemoteDescription(offer, function (done) {
-
-      expect(peer2.remoteDescription).to.equal(offer);
-      done();
-
-    }, function (error) {
-      throw error;
-    });
-  });
-
-  it('RTCPeerConnection.setRemoteDescription :: method < When > RTCPeerConnection.setRemoteDescription(invalid parameters)', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(function () {
-      peer1.setRemoteDescription(function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.setRemoteDescription(function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer1.setRemoteDescription(offer, function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.setRemoteDescription(offer, function () {});
-    }).to.throw(Error);
-
-    done();
-  });
-
-  it('RTCPeerConnection.createAnswer :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.createAnswer, 'function');
-    assert.equal(typeof peer2.createAnswer, 'function');
-
-    peer1.addStream(stream);
-
-    peer1.createOffer(function (offer) {
-      peer2.setRemoteDescription(offer);
-      peer2.createAnswer(function (answer) {
-
-        expect(answer.type).to.equal('answer');
-        assert.typeOf(answer.sdp, 'string');
-
-        expect(answer.sdp).to.contain('=audio');
-        expect(answer.sdp).to.contain('=video');
-        // expect(sdp.sdp).to.contain('=application'); // Note(J-O): I don't know what this is, commented
-
-        done();
-
-      }, function (error) {
-        throw error;
-      });
-    }, function(error) {
-      throw error;
-    });
-
-  });
-
-  it.skip('RTCPeerConnection.createAnswer :: method < When > RTCPeerConnection.createAnswer(invalid parameters)', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(function () {
-      peer1.createAnswer(function () {});
-    }).to.throw(err);
-
-    expect(function () {
-      peer2.createAnswer(function () {});
-    }).to.throw(err);
-
-    done();
-  });
-
-  it('RTCPeerConnection.setLocalDescription(offer) :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.setLocalDescription, 'function');
-
-    peer1.createOffer(function(offer) {
-      peer1.setLocalDescription(offer, function () {
-
-        expect(peer1.localDescription).to.deep.equal(offer);
-        done();
-
-      }, function (error) {
-        throw error;
-      });
-    }, function (error) {
-      throw error;
-    });
-  });
-
-  it.skip('RTCPeerConnection.setLocalDescription :: method < When > RTCPeerConnection.setLocalDescription(invalid parameters)', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(function () {
-      peer1.setLocalDescription(function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.setLocalDescription(function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer1.setLocalDescription(offer, function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.setLocalDescription(offer, function () {});
-    }).to.throw(Error);
-
-    done();
-  });
-
-  it('RTCPeerConnection.setLocalDescription(answer) :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer2.setLocalDescription, 'function');
-
-    peer1.createOffer(function(offer) {
-      peer2.setRemoteDescription(offer);
-      peer2.createAnswer(function(answer) {
-        peer2.setLocalDescription(answer, function () {
-
-          expect(peer2.localDescription).to.deep.equal(answer);
-          done();
-
-        }, function (error) {
-          throw error;
-        });
-        
-      });
-    });
-
-  });
-
-  it('RTCPeerConnection.setRemoteDescription(answer) :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.setRemoteDescription, 'function');
-
-    peer1.createOffer(function(offer) {
-      peer1.setLocalDescription(offer);
-      peer2.setRemoteDescription(offer);
-      peer2.createAnswer(function(answer) {
-        peer1.setRemoteDescription(answer, function () {
-
-          expect(peer1.remoteDescription).to.deep.equal(answer);
-          done();
-
-        }, function (error) {
-          throw error;
-        });
-        
-      });
-    });
-  });
-
-  it('RTCPeerConnection.addIceCandidate :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.addIceCandidate, 'function');
-    assert.equal(typeof peer2.addIceCandidate, 'function');
-
-    var peer1IceCount = 0;
-    var peer2IceCount = 0;
-    var peer1IceGathered = false;
-    var peer2IceGathered = false;
-    var iceAddSuccess1 = 0;
-    var iceAddSuccess2 = 0;
-
-    checkdone = function() {
-      if (peer1IceGathered && peer2IceGathered) {
-        assert.notEqual(peer1IceCount, 0);
-        assert.notEqual(peer2IceCount, 0);
-        assert.equal(peer2IceCount, iceAddSuccess1);
-        assert.equal(peer1IceCount, iceAddSuccess2);
-        done();
-      }
-    };
-
-    peer1.onicecandidate = function (event) {
-      var candidate = event.candidate;
-
-      if (candidate === null) {
-        peer1IceGathered = true;
-        checkdone();
-      } else {
-        ++peer1IceCount;
-        // TODO(J-O): since we don't test addIceCandidate here, can we remove it ?
-        peer2.addIceCandidate(candidate, function () {
-          ++iceAddSuccess2;
-        }, function (error) {
-          throw error;
-        });
-      }
-    };
-
-    peer2.onicecandidate = function (event) {
-      var candidate = event.candidate;
-
-      if (candidate === null) {
-        peer2IceGathered = true;
-        checkdone();
-      } else {
-        ++peer2IceCount;
-        // TODO(J-O): since we don't test addIceCandidate here, can we remove it ?
-        peer1.addIceCandidate(candidate, function () {
-          ++iceAddSuccess1;
-        }, function (error) {
-          throw error;
-        });
-      }
-    };
-
-    connect(peer1, peer2);
-  });
-
-  it('RTCPeerConnection.addIceCandidate :: method < When > RTCPeerConnection.addIceCandidate(invalid parameters)', function (done) {
-    this.timeout(testItemTimeout);
-
-    expect(function () {
-      peer1.addIceCandidate(candidates1[0]);
-    }).to.not.throw(Error);
-
-    expect(function () {
-      peer2.addIceCandidate(candidates2[0]);
-    }).to.not.throw(Error);
-
-    expect(function () {
-      peer1.addIceCandidate(candidates1[0], function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.addIceCandidate(candidates2[0], function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer1.addIceCandidate(candidates1[0], function () {}, function () {});
-    }).to.throw(Error);
-
-    expect(function () {
-      peer2.addIceCandidate(candidates2[0], function () {}, function () {});
-    }).to.throw(Error);
-
-    done();
-  });
-
-  it('RTCPeerConnection.close :: method', function (done) {
-    this.timeout(testItemTimeout);
-
-    assert.equal(typeof peer1.close, 'function');
-    assert.equal(typeof peer2.close, 'function');
-
-    peer1.close();
-    peer2.close();
-
-    setTimeout(function () {
-      expect(peer1.iceConnectionState).to.equal('close');
-      expect(peer2.iceConnectionState).to.equal('close');
-      expect(peer1.signalingState).to.equal('close');
-      expect(peer2.signalingState).to.equal('close');
-
-      expect(function () {
-        peer1.removeStream(stream);
-      }).to.throw(Error);
-
-      expect(function () {
-        peer2.removeStream(stream);
-      }).to.throw(Error);
-
-      expect(function () {
-        peer1.addStream(stream);
-      }).to.throw(Error);
-
-      expect(function () {
-        peer2.addStream(stream);
-      }).to.throw(Error);
-
-      expect(function () {
-        peer1.createOffer(function () {}, function () {});
-      }).to.throw(Error);
-
-      expect(function () {
-        peer2.createOffer(function () {}, function () {});
-      }).to.throw(Error);
-
-      expect(function () {
-        peer1.createAnswer(function () {}, function () {});
-      }).to.throw(Error);
-
-      expect(function () {
-        peer2.createAnswer(function () {}, function () {});
-      }).to.throw(Error);
-
-      done();
-    }, 2000);
-  });
-});
